+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. Version number will be decided later -->
+- feat: Add simple `classes` helper to filter + concatenate class-names
 
 ## 4.3.0 – 4.3.4